Attach aws views to aws components
