+++ src/views/AwsProvisioner/ViewRecentErrors/index.jsx
@@ -5,7 +5,7 @@ import ErrorPanel from '@mozilla-frontend-infra/components/ErrorPanel';
 import Spinner from '@mozilla-frontend-infra/components/Spinner';
 import recentErrors from './recentErrors.graphql';
 import Dashboard from '../../../components/Dashboard';
-import AwsRecentErrorsTable from '../../../components/AwsRecentErrorsTable';
+import AwsProvisionerErrorsTable from '../../../components/AwsProvisionerErrorsTable';
 
 @hot(module)
 @graphql(recentErrors)
@@ -23,7 +23,7 @@ export default class ViewRecentErrors extends Component {
             <ErrorPanel error={error.graphQLErrors[0].message} />
           )}
         {awsProvisionerRecentErrors && (
-          <AwsRecentErrorsTable recentErrors={awsProvisionerRecentErrors} />
+          <AwsProvisionerErrorsTable errors={awsProvisionerRecentErrors} />
         )}
       </Dashboard>
     );
