+++ src/views/AwsProvisioner/ViewWorkerTypeDefinition/deleteAwsProvisionerWorkerType.graphql
@@ -1,0 +1,3 @@
+mutation DeleteAwsProvisionerWorkerType($workerType: String!) {
+  deleteAwsProvisionerWorkerType(workerType: $workerType)
+}
